refactor: reorganize imports and configs in runtime_env module
